move itinerary logic to fragment from activity
