+++ traveler_ui_kit/src/main/java/com/guestlogix/traveleruikit/itinerary/ItineraryEmptyFragment.java
@@ -1,0 +1,21 @@
+package com.guestlogix.traveleruikit.itinerary;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.guestlogix.traveleruikit.R;
+
+public class ItineraryEmptyFragment extends Fragment {
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_itinerary_empty, container, false);
+    }
+}
