+++ traveler_ui_kit/src/main/res/layout/activity_itinerary.xml
@@ -1,23 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        tools:context="com.guestlogix.traveleruikit.activities.OrdersActivity">
-
-    <Button
-            android:id="@+id/btnFilters"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="right"
-            android:layout_margin="10dp"
-            android:text="Filters" />
-
-    <FrameLayout
-            android:id="@+id/itineraryContainer"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1" />
-</LinearLayout>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/itinerary_fragment_container"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent" />
