+++ traveler_ui_kit/src/main/res/values/strings.xml
@@ -132,5 +132,6 @@
     <string name="label_terms_and_conditions">Terms and conditions</string>
     <string name="itinerary_terms_and_conditions">Terms and conditions</string>
 
+    <string name="empty_itinerary">Your itinerary is empty</string>
 
 </resources>
